split integration and unit tests
