+++ examples/fastify-modules/__integration-test__/fastify-modules.spec.ts
