+++ examples/koa/tsconfig.json
@@ -9,5 +9,5 @@
     "skipLibCheck": true
   },
   "include": ["src"],
-  "exclude": ["node_modules", "dist", "test"]
+  "exclude": ["node_modules", "dist", "__integration-tests__"]
 }
