+++ examples/node-esm/jest.config.js
@@ -1,3 +1,6 @@
 export default {
   transform: {},
+  testMatch: [
+    '<rootDir>/**/__integration-tests__/**/?(*.)+(spec|test).m[jt]s?(x)',
+  ],
 }
