+++ examples/sveltekit/__integration-tests__/sveltekit.spec.ts
@@ -26,7 +26,9 @@ describe('SvelteKit integration', () => {
 			// Kill the port if it's used!
 			try {
 				execSync('fuser -k 3007/tcp');
-			} catch (error) {}
+			} catch (error) {
+				// console.error(error)
+			}
 
 			// Build svelteKit
 			execSync('yarn workspace example-sveltekit build');
@@ -86,10 +88,10 @@ describe('SvelteKit integration', () => {
 			let strIntro = '';
 			try {
 				// A-1/ Wait for the introspection query result getting our type "hello"
-				let resIntro = await page.waitForResponse((res) => res.url().endsWith('/api/graphql'), {
+				const resIntro = await page.waitForResponse((res) => res.url().endsWith('/api/graphql'), {
 					timeout: timings.waitForResponse
 				});
-				let jsonIntro = await resIntro.json();
+				const jsonIntro = await resIntro.json();
 				strIntro = JSON.stringify(jsonIntro, null, 0);
 			} catch (error) {
 				// We had an issue grabbing the introspection query result!
@@ -99,7 +101,7 @@ describe('SvelteKit integration', () => {
 
 				// B/ Check that GraphiQL is showing
 				expect(bodyContent).toContain(
-					`renderYogaGraphiQL(root,{\"endpoint\":\"/api/graphql\",\"defaultQuery\":\"query Hello {\\n\\thello\\n}\"})`
+					`renderYogaGraphiQL(root,{"endpoint":"/api/graphql","defaultQuery":"query Hello {\\n\\thello\\n}"})`
 				);
 			}
