+++ jest.config.js
@@ -8,19 +8,24 @@ const tsconfig = require(TSCONFIG)
 
 process.env.LC_ALL = 'en_US'
 
-const testMatch = ['<rootDir>/packages/**/?(*.)+(spec|test).[jt]s?(x)']
+const testMatch = []
 
 if (process.env.INTEGRATION_TEST === 'true') {
-  testMatch.pop()
   testMatch.push(
-    '<rootDir>/examples/**/?(*.)+(spec|test).[jt]s?(x)',
-    '!**/examples/node-esm/**',
+    '<rootDir>/**/__integration-tests__/**/?(*.)+(spec|test).[jt]s?(x)',
   )
   if (parseInt(process.versions.node.split('.')[0]) <= 14) {
-    testMatch.push('!**/examples/sveltekit/**')
-    testMatch.push('!**/examples/fastify*/**')
+    testMatch.push('!<rootDir>/**/examples/sveltekit/**')
+    testMatch.push('!<rootDir>/**/examples/fastify*/**')
   }
+} else {
+  testMatch.push(
+    '<rootDir>/packages/**/?(*.)+(spec|test).[jt]s?(x)',
+    '!**/__integration-tests__/**',
+  )
 }
+
+testMatch.push('!**/dist/**', '!**/.bob/**')
 
 module.exports = {
   testEnvironment: 'node',
